qr_service/requirements.txt
fastapi
uvicorn
qrcode
Pillow
cairosvg
